host 插件适配。
