+++ replugin-host-gradle/src/main/groovy/com/qihoo360/replugin/gradle/compat/CompatUtil.groovy
@@ -1,0 +1,49 @@
+/*
+ * Copyright (C) 2005-2017 Qihoo 360 Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed To in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ *
+ */
+
+package com.qihoo360.replugin.gradle.compat
+
+/**
+ * @author RePlugin Team
+ */
+class CompatUtil {
+    /**
+     * Android Gradle Plugin 是否大于等于 4.1.0
+     * @return
+     */
+    static def isGeAGP410() {
+        return (getAndroidGradlePluginVersionCompat() >= '4.1.0')
+    }
+
+    /**
+     * 获取 AGP 版本号
+     * @return 版本号字符串
+     */
+    static def getAndroidGradlePluginVersionCompat() {
+        String version = null
+        try {
+            Class versionModel = Class.forName("com.android.builder.model.Version")
+            def versionFiled = versionModel.getDeclaredField("ANDROID_GRADLE_PLUGIN_VERSION")
+            versionFiled.setAccessible(true)
+            version = versionFiled.get(null)
+        } catch (Exception e) {
+        }
+        return version
+    }
+
+    private CompatUtil() {}
+}
